+++ js/proto/reference_pb.d.ts
@@ -11,26 +11,20 @@ export class Ref extends jspb.Message {
     getAddress_asU8(): Uint8Array;
     getAddress_asB64(): string;
     setAddress(value: Uint8Array | string): Ref;
-
     getSecretkey(): Uint8Array | string;
     getSecretkey_asU8(): Uint8Array;
     getSecretkey_asB64(): string;
     setSecretkey(value: Uint8Array | string): Ref;
-
     getSalt(): Uint8Array | string;
     getSalt_asU8(): Uint8Array;
     getSalt_asB64(): string;
     setSalt(value: Uint8Array | string): Ref;
-
     getVersion(): number;
     setVersion(value: number): Ref;
-
     getType(): Ref.RefType;
     setType(value: Ref.RefType): Ref;
-
     getSize(): number;
     setSize(value: number): Ref;
-
 
     serializeBinary(): Uint8Array;
     toObject(includeInstance?: boolean): Ref.AsObject;
@@ -65,12 +59,10 @@ export class RefsWithNonce extends jspb.Message {
     getRefsList(): Array<Ref>;
     setRefsList(value: Array<Ref>): RefsWithNonce;
     addRefs(value?: Ref, index?: number): Ref;
-
     getNonce(): Uint8Array | string;
     getNonce_asU8(): Uint8Array;
     getNonce_asB64(): string;
     setNonce(value: Uint8Array | string): RefsWithNonce;
-
 
     serializeBinary(): Uint8Array;
     toObject(includeInstance?: boolean): RefsWithNonce.AsObject;
@@ -86,5 +78,39 @@ export namespace RefsWithNonce {
     export type AsObject = {
         refsList: Array<Ref.AsObject>,
         nonce: Uint8Array | string,
+    }
+}
+
+export class Link extends jspb.Message { 
+
+    hasHeader(): boolean;
+    clearHeader(): void;
+    getHeader(): Ref | undefined;
+    setHeader(value?: Ref): Link;
+    clearBodyList(): void;
+    getBodyList(): Array<Ref>;
+    setBodyList(value: Array<Ref>): Link;
+    addBody(value?: Ref, index?: number): Ref;
+
+    hasTrailer(): boolean;
+    clearTrailer(): void;
+    getTrailer(): Ref | undefined;
+    setTrailer(value?: Ref): Link;
+
+    serializeBinary(): Uint8Array;
+    toObject(includeInstance?: boolean): Link.AsObject;
+    static toObject(includeInstance: boolean, msg: Link): Link.AsObject;
+    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
+    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
+    static serializeBinaryToWriter(message: Link, writer: jspb.BinaryWriter): void;
+    static deserializeBinary(bytes: Uint8Array): Link;
+    static deserializeBinaryFromReader(message: Link, reader: jspb.BinaryReader): Link;
+}
+
+export namespace Link {
+    export type AsObject = {
+        header?: Ref.AsObject,
+        bodyList: Array<Ref.AsObject>,
+        trailer?: Ref.AsObject,
     }
 }